+++ CHANGES.rst
@@ -1,7 +1,8 @@
 0.1.1
 -----
 * Bug fix: IStaticResource fails to look up self.request if nested 2-deep
-* Bug fix: name collisions with version_helper.py
+* Bug fix: Name collisions with version_helper.py
+* Bug fix: Subpath glob matching always respects case
 
 0.1.0
 -----